Bug fixes in zynseq.

+++ zynlibs/zynseq/sequence.h
@@ -249,7 +249,7 @@ class Sequence {
     uint16_t m_nTimeSig = 4;                    // Current time signature (beats in bar)
     uint32_t m_nId;                             // Sequence id (sequence number)
     uint8_t m_nFollowAction = FOLLOW_ACTION_LOOP; // Sequence follow action
-    uint8_t m_nFollowParam = 0;                 // Parameter for follow action, e.g. jump offset
+    uint32_t m_nFollowParam = 0;                // Parameter for follow action, e.g. jump offset
     uint8_t m_nState = STOPPED;                 // Play state of sequence
     uint8_t m_nMode = 1;                        // Bitwise flags: stop mode (bits 0..1), start mode (bit 2).
     uint8_t m_nGroup = 0;                       // Sequence's mutually exclusive group
